feat: implement switch append-only and delete-load
- Change Dag Schedular to hourly
- Make Logs Message Dynamic

+++ plugins/operators/load_dimension.py
@@ -15,25 +15,25 @@ class LoadDimensionOperator(BaseOperator):
                  table="",
                  redshift_conn_id="",
                  select_query="",
+                 append_data = False,
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
-        # Map params here
-        # Example:
         self.table = table
         self.select_query = select_query
         self.redshift_conn_id = redshift_conn_id
+        self.append_data = append_data
 
     def execute(self, context):
-        self.log.info('Start LoadDimensionOperator')
+        self.log.info(f"Start LoadDimensionOperator for {self.table}")
 
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
 
-        self.log.info("Clearing data from destination Redshift table")
-        redshift.run("DELETE FROM {}".format(self.table))
+        if not self.append_data:
+            self.log.info("Clearing data from destination Redshift table")
+            redshift.run("DELETE FROM {}".format(self.table))
 
         self.log.info("Copying data from Staging table to Dimension table")
-
         load_data_sql = LoadDimensionOperator.insert_sql.format(
             self.table,
             self.select_query
