+++ plugins/operators/load_fact.py
@@ -16,25 +16,25 @@ class LoadFactOperator(BaseOperator):
                  table="",
                  redshift_conn_id="",
                  select_query="",
+                 append_data = False,
                  *args, **kwargs):
 
         super(LoadFactOperator, self).__init__(*args, **kwargs)
-        # Map params here
-        # Example:
         self.table = table
         self.select_query = select_query
         self.redshift_conn_id = redshift_conn_id
+        self.append_data = append_data
 
     def execute(self, context):
-        self.log.info('Start LoadFactOperator')
+        self.log.info(f"Start LoadFactOperator for {self.table}")
 
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
 
-        self.log.info("Clearing data from destination Redshift table")
-        redshift.run("DELETE FROM {}".format(self.table))
+        if not self.append_data:
+            self.log.info("Clearing data from Destination Redshift table")
+            redshift.run("DELETE FROM {}".format(self.table))
 
         self.log.info("Copying data from Staging table to Fact table")
-
         load_data_sql = LoadFactOperator.insert_sql.format(
             self.table,
             self.select_query
